refactor(sanity): add useVersionOperations, replace createVersion on dialog and chip
